+++ src/lib/comms/include/sol-mqtt.h
@@ -343,7 +343,7 @@ struct sol_mqtt_config {
  *
  * @return New mqtt object on sucess, NULL otherwise
  */
-struct sol_mqtt *sol_mqtt_connect(const char *host, int port, const struct sol_mqtt_config *config, void *data);
+struct sol_mqtt *sol_mqtt_connect(const char *host, uint16_t port, const struct sol_mqtt_config *config, void *data);
 
 /**
  * @brief Reestablish the connection to the MQTT broker
